Rename module to cached_module
