+++ packages/vm/src/modules/cached_module.rs
