+++ packages/vm/src/modules/mod.rs
@@ -1,11 +1,11 @@
+mod cached_module;
 mod file_system_cache;
 mod in_memory_cache;
 mod pinned_memory_cache;
-mod sized_module;
 mod versioning;
 
+pub use cached_module::CachedModule;
 pub use file_system_cache::{FileSystemCache, NewFileSystemCacheError};
 pub use in_memory_cache::InMemoryCache;
 pub use pinned_memory_cache::PinnedMemoryCache;
-pub use sized_module::CachedModule;
 pub use versioning::current_wasmer_module_version;
